+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Fix typo in error message [Bradley](https://github.com/bradleyisotope).
 * Fix excon adapter to properly handle queries specified as a hash.
+* Fix excon adapter to stub a response with a hash as excon expects
+  [Wesley Beary](https://github.com/geemus).
 
 ## 1.10.0 (May 18, 2011)